--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,7 +16,7 @@ name = "fuse-query"
 path = "src/bin/fuse-query.rs"
 
 [dependencies]
-arrow = { git= "https://github.com/apache/arrow"}
+arrow = { git = "https://github.com/apache/arrow", rev = "0c419fc"}
 async-trait = "0.1"
 bincode = "1.3.1"
 futures = "0.3"