--- a/benches/bench_main.rs
+++ b/benches/bench_main.rs
@@ -8,4 +8,5 @@ mod benchmarks;
 
 criterion_main! {
     benchmarks::bench_pipeline::benches,
+    benchmarks::bench_aggregate::benches,
 }