--- a/benches/benchmarks/mod.rs
+++ b/benches/benchmarks/mod.rs
@@ -3,3 +3,4 @@
 // Code is licensed under AGPL License, Version 3.0.
 
 pub mod bench_pipeline;
+pub mod bench_aggregate;