--- a/benches/benchmarks/mod.rs
+++ b/benches/benchmarks/mod.rs
@@ -2,5 +2,5 @@
 //
 // Code is licensed under AGPL License, Version 3.0.
 
-pub mod bench_pipeline;
 pub mod bench_aggregate;
+pub mod bench_pipeline;