--- a/src/contexts/options.rs
+++ b/src/contexts/options.rs
@@ -15,7 +15,11 @@ impl Options {
         let settings = SettingMap::create();
         settings.try_set_string("log_level", "debug".to_string(), "Log level")?;
         settings.try_set_u64("num_cpus", num_cpus::get() as u64, "The numbers of the pc")?;
-        settings.try_set_string("mysql_listen_host", "127.0.0.1".to_string(), "MySQL server bind host")?;
+        settings.try_set_string(
+            "mysql_listen_host",
+            "127.0.0.1".to_string(),
+            "MySQL server bind host",
+        )?;
         settings.try_set_u64("mysql_handler_port", 3307, "MySQL protocol port")?;
         settings.try_set_u64(
             "mysql_handler_thread_num",