--- a/src/servers/mysql/mysql_handler.rs
+++ b/src/servers/mysql/mysql_handler.rs
@@ -132,8 +132,11 @@ impl MySQLHandler {
     }
 
     pub fn start(&self) -> FuseQueryResult<()> {
-        let listener =
-            net::TcpListener::bind(format!("{}:{}", self.opts.get_mysql_listen_host()?, self.opts.get_mysql_handler_port()?))?;
+        let listener = net::TcpListener::bind(format!(
+            "{}:{}",
+            self.opts.get_mysql_listen_host()?,
+            self.opts.get_mysql_handler_port()?
+        ))?;
         let pool = ThreadPool::new(self.opts.get_mysql_handler_thread_num()? as usize);
 
         for stream in listener.incoming() {