## Changes:

1. Improve numbers_mt generate source performance (About 2x improvement).
2. Introduce settings `mysql_listen_host`, defaults to `12.0.0.1` for security reason.
3. Set `max_block_size` to `65536` for better performance.

## Performance comparison 

- fuse-query


```
mysql> SELECT count(number) FROM system.numbers_mt(10000000000);
+---------------+
| count(number) |
+---------------+
|   10000000000 |
+---------------+
1 row in set (0.28 sec)

mysql> SELECT max(number) FROM system.numbers_mt(10000000000);
+-------------+
| max(number) |
+-------------+
|  9999999999 |
+-------------+
1 row in set (0.45 sec)

mysql> SELECT avg(number) FROM system.numbers_mt(10000000000);
+--------------+
| avg(number)  |
+--------------+
| 4999999999.5 |
+--------------+
1 row in set (0.50 sec)
```


- ClickHouse 

```
:)  SELECT count(number) FROM  numbers_mt(10000000000)

SELECT count(number)
FROM numbers_mt(10000000000)


1 rows in set. Elapsed: 0.520 sec. Processed 10.00 billion rows, 80.00 GB (19.24 billion rows/s., 153.94 GB/s.)

:) SELECT max(number) FROM  numbers_mt(10000000000);

SELECT max(number)
FROM numbers_mt(10000000000)


1 rows in set. Elapsed: 1.525 sec. Processed 10.00 billion rows, 80.00 GB (6.56 billion rows/s., 52.46 GB/s.)

:) SELECT avg(number) FROM  numbers_mt(10000000000);

SELECT avg(number)
FROM numbers_mt(10000000000)


1 rows in set. Elapsed: 1.167 sec. Processed 10.00 billion rows, 80.00 GB (8.57 billion rows/s., 68.56 GB/s.)

```