@sundy-li 2021-01-16
# How I did this

## Background

There are two SQL queries, which do aggregation `count` and `max`,  but their performances are close.
Theoretically, `count` query should performances better than `max`, because `count` just call length of array, but the other involves `min_max_helper` vectorized computing in `arrow`.

## Profile and tunning


- Count, sum(top3) = 86.44%
```
    31.70%  tokio-runtime-w  libc-2.19.so        [.] __memcpy_sse2_unaligned
    27.95%  tokio-runtime-w  libc-2.19.so        [.] memset
    26.79%  tokio-runtime-w  fuse-query          [.] <alloc::vec::Vec<T> as alloc::vec::spec_from_iter::SpecFromIter<T,I>>::from_iter
     1.69%  tokio-runtime-w  libc-2.19.so        [.] malloc
```

- Max, sum(top4) = 87.31%
```
    28.34%  tokio-runtime-w  libc-2.19.so       [.] __memcpy_sse2_unaligned
    24.52%  tokio-runtime-w  fuse-query         [.] <alloc::vec::Vec<T> as alloc::vec::spec_from_iter::SpecFromIter<T,I>>::from_iter
    24.20%  tokio-runtime-w  libc-2.19.so       [.] memset
    10.25%  tokio-runtime-w  fuse-query         [.] arrow::compute::kernels::aggregate::max
     1.60
```

As we can see, `max` query has about `10.25%` extra cpu costs in `arrow::compute::kernels::aggregate::max`.
But why it performances the same as the `count` query?

After some thinking, I realized that must be the problem of source data generation, the speed of upstream generation is slow, so the downstream are hungry.  Just like two men are waiting for food, one of them can finish the dinner in 5 minutes, the other may take 1 minute, but food preparation time may take an hour, so we see they just finish the dinner at the same time.


###  Why they generate the data too slow? 

The top  `memcpy` and `alloc`  seem to be reasonable if we generate large vectors.
Look at the clickhouse perf result, it can achieve `141.59 GB/s` in the `count` query, and it takes `60.78` in generate the source data, which is almost the double of `fuse-query`

```
    --60.78%--DB::ISource::work
                |
                 --60.66%--DB::(anonymous namespace)::NumbersMultiThreadedSource::generate
                           |
                           |--3.00%--DB::SourceWithProgress::progress
                           |          |
                           |           --1.10%--DB::EnabledQuota::used
```


 It's hard to find the problem by watching the profile without experiments.


### Let's do some experiments

- I did a benchmark test in [intel mlc](https://software.intel.com/content/www/us/en/develop/articles/intelr-memory-latency-checker.html)

It shows we can get `115GB/s`
```
ALL Reads        :  115828.6
```

- Let's benchmark the memory bandwidth to see how much performance can we got in theoretically.
I create a sample code in [here](), it shows we can get `165GB/s` in blocksize `65536` in my PC (32Core 256G)

```
> ./target/release/examples/m1 32 65536
Memory Bandwidth: 165 GB/s, cost: 484.745264ms
```
Seems the result of `mlc` is Untrustworthy. But `fuse-query` just reach `50GB/s~60GB/s`. What is exciting is `fuse-query` can have a lot of space for optimization.

### Tunning

I did a perf test for sample code, it shows `memcpy` is not the bottleneck.  So I suddenly realize the difference between them are `UInt64Array::from` of `arrow`, then I catch this problem, it's the issue about `UInt64Array::from`, it uses `memcpy` to copy bytes from vector to generate the buffer.


### Solve

As the code shows, I use some unsafe tricks to avoid memcpy. And we can achieve the performance about ``