@alexey-milovidov 2021-01-16
> Seems the result of mlc is Untrustworthy.

It looks like you are comparing main memory vs. cache performance.
The query from `system.numbers` can generate and process data entirely in cache, without resolving to main memory.