@sundy-li 2021-01-16
> The query from `system.numbers` can generate and process data entirely in cache, without resolving to main memory.

I see, thank you, Alexey.