@l1t1 2021-01-28
clickhouse desnt support ` system.numbers_mt(1000)`
Syntax error: failed at position 42 ('('):

SELECT max(number) FROM system.numbers_mt(10000000000);

Unmatched parentheses: (